# Cache construction of middleware handlers