--- a/CHANGES/9158.misc.rst
+++ b/CHANGES/9158.misc.rst
@@ -1,0 +1,3 @@
+Significantly improved performance of middlewares -- by :user:`bdraco`.
+
+The construction of the middleware wrappers is now cached and is built once per handler instead of on every request.