@Dreamsorcerer 2024-09-16
> Doesnt the cache need to be invalidated if an additional middleware is added at some late stage? (Know too little about internals to know if this can happen)

They're frozen, they can't be modified once the app is started.