@bdraco 2024-09-17
Need to look at the lru stats once I have stable internet